add modules CommonJS, dont't work reviews
