+++ src/index.html
@@ -127,9 +127,6 @@
   </article>
 </template>
 
-<script src="js/form.js"></script>
-<script src="js/game.js"></script>
-<script src="js/reviews.js"></script>
 <script src="js/main.js"></script>
 
 </body>
