+++ src/js/load.js
@@ -1,0 +1,17 @@
+'use strict';
+
+var reviews = [];
+
+module.exports = function(url, callback, callbackName) {
+  if (!callbackName) {
+    callbackName = 'cb' + Date.now();
+  }
+  window[callbackName] = function(data) {
+    reviews = data;
+    callback(data);
+  };
+
+  var script = document.createElement('script');
+  script.src = url + '?callback=' + callbackName;
+  document.body.appendChild(script);
+};
